Remove Postgres conversion from diff structs

+++ daiklave-core/src/merits/update.rs
@@ -9,7 +9,7 @@ use std::collections::HashSet;
 #[derive(Debug, Default)]
 pub struct MeritDiff {
     insert_merit_templates: Vec<Merit>,
-    insert_merit_instance: Vec<(i32, i16, Option<String>)>,
+    insert_merit_instance: Vec<(i32, u8, Option<String>)>,
     remove_merit_instances: Vec<i32>,
 }
 
@@ -33,7 +33,7 @@ pub fn compare_merits(old_merits: &[Merit], new_merits: &[Merit]) -> MeritDiff {
         } else if merit.instance_id().is_placeholder() {
             diff.insert_merit_instance.push((
                 *merit.template_id(),
-                merit.dots() as i16,
+                merit.dots(),
                 merit.detail().map(|s| s.to_owned()),
             ))
         } else {
